+++ migrations/007-users_groups.sql
@@ -3,7 +3,7 @@ CREATE TABLE "users_groups" (
 	"id" serial NOT NULL,
 	"user_id" int NOT NULL,
 	"group_id" int NOT NULL,
-	"created_at" TIMESTAMPTZ,
+	"created_at" TIMESTAMPTZ NOT NULL,
 	PRIMARY KEY("id")
 );
 -- +migrate Down