+++ repositories/functionalities.go
@@ -2,9 +2,18 @@ package repositories
 
 import (
 	"regexp"
+	g "service/global"
+	"service/static_models"
+	"time"
 
+	"github.com/dgrijalva/jwt-go"
 	"golang.org/x/crypto/bcrypt"
 )
+
+func hasPassword(password string) string {
+	bytes, _ := bcrypt.GenerateFromPassword([]byte(password), 16)
+	return string(bytes)
+}
 
 func (u *User) IsHashed() bool {
 	// Bcrypt hash pattern
@@ -16,6 +25,43 @@ func (u *User) IsHashed() bool {
 }
 
 func (u *User) HashPassword() {
-	bytes, _ := bcrypt.GenerateFromPassword([]byte(u.Password), 16)
-	u.Password = string(bytes)
+	u.Password = hasPassword(u.Password)
+}
+
+func (u *User) IsSamePassword(rawPassword string) bool {
+	if err := bcrypt.CompareHashAndPassword([]byte(u.Password), []byte(rawPassword)); err != nil {
+		return false
+	}
+
+	return true
+}
+
+func (u *User) GenerateToken(refreshToken ...bool) Token {
+	var isRefreshToken = false
+	if len(refreshToken) > 0 {
+		isRefreshToken = refreshToken[0]
+	}
+
+	var period = g.CFG.AccessTokenLifePeriod
+	if isRefreshToken {
+		period = g.CFG.RefreshTokenLifePeriod
+	}
+	expirationTime := time.Now().Add(time.Duration(period) * (time.Hour * 24))
+
+	var tokenType = static_models.AccessTokenType
+	if isRefreshToken {
+		tokenType = static_models.RefreshTokenType
+	}
+	claims := &static_models.Claims{
+		UserId: u.ID,
+		Type:   tokenType,
+		StandardClaims: jwt.StandardClaims{
+			ExpiresAt: expirationTime.Unix(),
+		},
+	}
+
+	tkn := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	tokenString, _ := tkn.SignedString(g.SecretKeyBytes)
+
+	return NewToken(tokenString, isRefreshToken, u.ID, expirationTime)
 }