+++ repositories/initiator.go
@@ -15,3 +15,14 @@ func NewUser[T any](instance *T) User {
 	}
 	return user
 }
+
+func NewToken(content string, refreshToken bool, userId int32, expiresAt time.Time) Token {
+	token := Token{
+		Token:          content,
+		IsRefreshToken: refreshToken,
+		UserID:         userId,
+		ExpiresAt:      expiresAt,
+		CreatedAt:      time.Now(),
+	}
+	return token
+}