+++ repositories/query.sql.go
@@ -9,17 +9,83 @@ import (
 	"context"
 	"database/sql"
 	"time"
-	"service/pkg/errors"
-	"service/global"
-	"service/i18n/i18n_interfaces"
 )
+
+const createAccessToken = `-- name: CreateAccessToken :one
+INSERT INTO tokens (
+  token, is_refresh_token, user_id, expires_at, created_at
+) VALUES (
+  $1, FALSE, $2, $3, $4
+)
+RETURNING id, token, is_refresh_token, user_id, expires_at, created_at
+`
+
+type CreateAccessTokenParams struct {
+	Token     string    `json:"token"`
+	UserID    int32     `json:"user_id"`
+	ExpiresAt time.Time `json:"expires_at"`
+	CreatedAt time.Time `json:"created_at"`
+}
+
+func (q *Queries) CreateAccessToken(ctx context.Context, arg CreateAccessTokenParams) (Token, error) {
+	row := q.db.QueryRowContext(ctx, createAccessToken,
+		arg.Token,
+		arg.UserID,
+		arg.ExpiresAt,
+		arg.CreatedAt,
+	)
+	var i Token
+	err := row.Scan(
+		&i.ID,
+		&i.Token,
+		&i.IsRefreshToken,
+		&i.UserID,
+		&i.ExpiresAt,
+		&i.CreatedAt,
+	)
+	return i, err
+}
+
+const createRefreshToken = `-- name: CreateRefreshToken :one
+INSERT INTO tokens (
+  token, is_refresh_token, user_id, expires_at, created_at
+) VALUES (
+  $1, TRUE, $2, $3, $4
+)
+RETURNING id, token, is_refresh_token, user_id, expires_at, created_at
+`
+
+type CreateRefreshTokenParams struct {
+	Token     string    `json:"token"`
+	UserID    int32     `json:"user_id"`
+	ExpiresAt time.Time `json:"expires_at"`
+	CreatedAt time.Time `json:"created_at"`
+}
+
+func (q *Queries) CreateRefreshToken(ctx context.Context, arg CreateRefreshTokenParams) (Token, error) {
+	row := q.db.QueryRowContext(ctx, createRefreshToken,
+		arg.Token,
+		arg.UserID,
+		arg.ExpiresAt,
+		arg.CreatedAt,
+	)
+	var i Token
+	err := row.Scan(
+		&i.ID,
+		&i.Token,
+		&i.IsRefreshToken,
+		&i.UserID,
+		&i.ExpiresAt,
+		&i.CreatedAt,
+	)
+	return i, err
+}
 
 const getUserById = `-- name: GetUserById :one
 SELECT id, phone_number, email, password, profile, first_name, last_name, display_name, gender, is_active, registered, deactivation_reason, is_admin, otp_remaining_attempts, otp_code, otp_due_date, is_superuser, created_at FROM users WHERE id = $1
 `
 
 func (q *Queries) GetUserById(ctx context.Context, id int32) (User, error) {
-	translator := ctx.Value(g.TranslatorKey).(i18n_interfaces.TranslatorI)
 	row := q.db.QueryRowContext(ctx, getUserById, id)
 	var i User
 	err := row.Scan(
@@ -42,9 +108,6 @@ func (q *Queries) GetUserById(ctx context.Context, id int32) (User, error) {
 		&i.IsSuperuser,
 		&i.CreatedAt,
 	)
-	if err != nil {
-		panic(errors.New(errors.UnexpectedStatus, translator.StatusCodes().InternalServerError(), err.Error()))
-	}
 	return i, err
 }
 
@@ -53,7 +116,6 @@ SELECT id, phone_number, email, password, profile, first_name, last_name, displa
 `
 
 func (q *Queries) LoginUserWithEmail(ctx context.Context, email sql.NullString) (User, error) {
-	translator := ctx.Value(g.TranslatorKey).(i18n_interfaces.TranslatorI)
 	row := q.db.QueryRowContext(ctx, loginUserWithEmail, email)
 	var i User
 	err := row.Scan(
@@ -76,9 +138,6 @@ func (q *Queries) LoginUserWithEmail(ctx context.Context, email sql.NullString) 
 		&i.IsSuperuser,
 		&i.CreatedAt,
 	)
-	if err != nil {
-		panic(errors.New(errors.UnexpectedStatus, translator.StatusCodes().InternalServerError(), err.Error()))
-	}
 	return i, err
 }
 
@@ -87,7 +146,6 @@ SELECT id, phone_number, email, password, profile, first_name, last_name, displa
 `
 
 func (q *Queries) LoginUserWithPhoneNumber(ctx context.Context, phoneNumber string) (User, error) {
-	translator := ctx.Value(g.TranslatorKey).(i18n_interfaces.TranslatorI)
 	row := q.db.QueryRowContext(ctx, loginUserWithPhoneNumber, phoneNumber)
 	var i User
 	err := row.Scan(
@@ -110,9 +168,6 @@ func (q *Queries) LoginUserWithPhoneNumber(ctx context.Context, phoneNumber stri
 		&i.IsSuperuser,
 		&i.CreatedAt,
 	)
-	if err != nil {
-		panic(errors.New(errors.UnexpectedStatus, translator.StatusCodes().InternalServerError(), err.Error()))
-	}
 	return i, err
 }
 
@@ -134,7 +189,6 @@ type RegisterUserParams struct {
 }
 
 func (q *Queries) RegisterUser(ctx context.Context, arg RegisterUserParams) (User, error) {
-	translator := ctx.Value(g.TranslatorKey).(i18n_interfaces.TranslatorI)
 	row := q.db.QueryRowContext(ctx, registerUser,
 		arg.PhoneNumber,
 		arg.Email,
@@ -163,8 +217,5 @@ func (q *Queries) RegisterUser(ctx context.Context, arg RegisterUserParams) (Use
 		&i.IsSuperuser,
 		&i.CreatedAt,
 	)
-	if err != nil {
-		panic(errors.New(errors.UnexpectedStatus, translator.StatusCodes().InternalServerError(), err.Error()))
-	}
 	return i, err
 }